Julian

+++ jobSearch_app/views.py
@@ -41,35 +41,6 @@ def find_jobs(request):
         Job.objects.create(job_title=pos, company=job['Company'], location=job['Location'], salary_min=job['salary_min'], salary_max=job['salary_max'], job_url=job['JobUrl'], job_desc=job['JobDesc'], summary=job['Summary'], post_date=job['PostDate'])
     return redirect("/job")
 
-# @validate_request
-# def like_job(request, logged_user, like_job_id):
-#     if request.method == "POST":
-#         logged_user = User.objects.get(id=logged_user)
-#         job_to_like = Job.objects.get(id=like_job_id)
-#         logged_user.job_likes.add(job_to_like)
-#         if job_to_like in logged_user.job_dislikes.all():
-#             logged_user.job_dislikes.remove(logged_user)
-#     return redirect("/job")
-
-# @validate_request
-# def dislike_job(request, logged_user, dislike_job_id):
-#     if request.method == "POST":
-#         logged_user = User.objects.get(id=logged_user)
-#         job_to_dislike = Job.objects.get(id=dislike_job_id)
-#         logged_user.dislikes.add(job_to_dislike)
-#         if job_to_dislike in logged_user.job_likes.all():
-#             logged_user.job_likes.remove(logged_user)
-        
-#     return redirect("/job")
-
-# @validate_request
-# def reset_job(request, logged_user, reset_job_id):
-#     if request.method == "POST":
-#         job_to_reset = Job.objects.get(id=reset_job_id)
-#         logged_user.job_likes.remove(job_to_reset)
-#         logged_user.job_dislikes.remove(job_to_reset)
-#     return redirect("/job")
-
 @validate_request
 def like(request, user):
     if request.method == "GET":
@@ -183,13 +154,24 @@ def create_note(request, job_id):
     if request.method == 'POST':
         this_job = Job.objects.get(id=job_id)
         this_user = User.objects.get(id=request.POST['user'])
-        this_note = Note.objects.create(creator= this_user, job_id=this_job, desc=request.POST['desc'])
+        Note.objects.create(creator= this_user, job_id=this_job, desc=request.POST['desc'])
         return redirect(f'/job/{job_id}')
     else: 
         return redirect('/')
 
+def note_edit(request, note_id):
+    if request.method == "POST":
+        this_note = Note.objects.filter(id=note_id)
+        if this_note:
+            this_note = this_note[0]
+            this_note.desc = request.POST['desc']
+            this_note.save()
+            return redirect(request.META.get('HTTP_REFERER', 'redirect_if_referer_not_found'))
+    return redirect(f"/job")
+
 def delete_note(request, job_id, note_id):
+    if request.method == "POST":
         this_note = Note.objects.get(id=note_id)
         this_note.delete()
-
         return redirect(f'/job/{job_id}')
+    return redirect("/job")
